added warnings

+++ src/app/booked/booked.component.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ApiservicesService } from '../services/apiservices.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-booked',
@@ -20,16 +21,33 @@ export class BookedComponent {
     this.bookings = storedBookings ? JSON.parse(storedBookings) : [];
   }
   deleteBooking(booking: any): void {
-    let stored = localStorage.getItem('bookings');
-    let bookings = stored ? JSON.parse(stored) : [];
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result: { isConfirmed: any; }) => {
+      if (result.isConfirmed) {
+        let stored = localStorage.getItem('bookings');
+        let bookings = stored ? JSON.parse(stored) : [];
 
-    let bookingId = booking.customerId || booking.id;
-    let index = bookings.findIndex((b: { customerId: any; id: any; }) => (b.customerId || b.id) === bookingId);
-    if (index !== -1) {
-      bookings.splice(index, 1);
-      localStorage.setItem('bookings', JSON.stringify(bookings));
-      this.bookings = bookings;
-    }
+        let bookingId = booking.customerId || booking.id;
+        let index = bookings.findIndex((b: { customerId: any; id: any; }) => (b.customerId || b.id) === bookingId);
+        if (index !== -1) {
+          bookings.splice(index, 1);
+          localStorage.setItem('bookings', JSON.stringify(bookings));
+          this.bookings = bookings;
+          Swal.fire(
+            'Deleted!',
+            'Your booking has been deleted.',
+            'success'
+          );
+        }
+      }
+    });
   }
   
 }
